Updates gatsby-browser to TS
