--- gatsby-browser.js
@@ -1,10 +1,0 @@
-exports.onRouteUpdate = () => {
-  const mobileMenuElements = document.querySelectorAll(
-    '[class*=StyledMobileMenu]'
-  );
-  mobileMenuElements.forEach((item) => {
-    if (item.classList.contains('is-open')) {
-      item.classList.remove('is-open');
-    }
-  });
-};
